implement filtered tools rendering in ToolsGrid component
